Revert "refactor(backoffice-v2): changed the conditions in which case and task actions are disabled (#812)"
This reverts commit a7a297d4cb7666a78149c4a1fd71bc5753e095f2.
