+++ apps/backoffice-v2/src/pages/Entity/components/KycBlock/hooks/useKycBlock/useKycBlock.tsx
@@ -7,7 +7,6 @@ import {
 import { octetToFileType } from '../../../../../../common/octet-to-file-type/octet-to-file-type';
 import { capitalize } from '../../../../../../common/utils/capitalize/capitalize';
 import { safeEvery } from '@ballerine/common';
-import { useCaseDecision } from '../../../Case/hooks/useCaseDecision/useCaseDecision';
 
 export const useKycBlock = ({
   parentWorkflowId,
@@ -16,7 +15,6 @@ export const useKycBlock = ({
   childWorkflow: TWorkflowById['childWorkflows'][number];
   parentWorkflowId: string;
 }) => {
-  const { noAction } = useCaseDecision();
   const docsData = useStorageFilesQuery(
     childWorkflow?.context?.documents?.flatMap(({ pages }) =>
       pages?.map(({ ballerineFileId }) => ballerineFileId),
@@ -131,9 +129,10 @@ export const useKycBlock = ({
         fileType: type,
       })) ?? [],
   );
-  const hasDecision =
-    safeEvery(childWorkflow?.context?.documents, document => !!document?.decision?.status) ||
-    noAction;
+  const hasDecision = safeEvery(
+    childWorkflow?.context?.documents,
+    document => !!document?.decision?.status,
+  );
 
   return [
     [
