+++ apps/backoffice-v2/src/pages/Entity/hooks/useTasks/useTasks.tsx
@@ -16,7 +16,6 @@ import { useMemo } from 'react';
 import { toStartCase } from '../../../../common/utils/to-start-case/to-start-case';
 
 import { octetToFileType } from '../../../../common/octet-to-file-type/octet-to-file-type';
-import { useCaseDecision } from '../../components/Case/hooks/useCaseDecision/useCaseDecision';
 
 export const useTasks = ({
   workflow,
@@ -38,7 +37,6 @@ export const useTasks = ({
       pages?.map(({ ballerineFileId }) => ballerineFileId),
     ),
   );
-  const { noAction } = useCaseDecision();
 
   const results: Array<Array<string>> = [];
   workflow?.context?.documents?.forEach((document, docIndex) => {
@@ -108,8 +106,7 @@ export const useTasks = ({
                           value: 'Reject',
                           data: {
                             id,
-                            disabled:
-                              (!isDoneWithRevision && Boolean(decision?.status)) || noAction,
+                            disabled: !isDoneWithRevision && Boolean(decision?.status),
                             decision: 'reject',
                           },
                         },
@@ -118,8 +115,7 @@ export const useTasks = ({
                           value: 'Approve',
                           data: {
                             id,
-                            disabled:
-                              (!isDoneWithRevision && Boolean(decision?.status)) || noAction,
+                            disabled: !isDoneWithRevision && Boolean(decision?.status),
                             decision: 'approve',
                           },
                         },
